move testing chain spec to own file
