+++ node/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod chain_spec;
+pub mod testing_chain_spec;
 pub mod client;
 pub mod rpc;
 pub mod service;
