+++ node/src/main.rs
@@ -2,6 +2,7 @@
 #![warn(missing_docs)]
 
 mod chain_spec;
+mod testing_chain_spec;
 #[macro_use]
 mod service;
 mod cli;
